elang/lir: Use |size_t| for |BasicBlockOperands::size()|.

+++ elang/lir/instructions.h
@@ -143,7 +143,7 @@ class ELANG_LIR_EXPORT BasicBlockOperands final {
   Iterator begin() const { return Iterator(start_); }
   bool empty() const { return start_ == end_; }
   Iterator end() const { return Iterator(end_); }
-  int size() const { return static_cast<int>(end_ - start_); }
+  size_t size() const { return end_ - start_; }
 
  private:
   friend class Instruction;
